requirements.txt
python-telegram-bot[job-queue]==21.6
smartapi-python==1.5.5
pyotp==2.9.0
requests==2.31.0
pandas==2.2.0
logzero==1.7.0
apscheduler==3.10.4
